code/web-server-object_detection/requirements.txt
Flask==1.0.2
gunicorn==19.9.0
numpy==1.16.3
tensorflow==1.13.1
opencv-python-headless
Pillow==6.0.0
Keras==2.2.4
